fix: close T.1 QA/runtime findings
