+++ crates/atm-daemon/src/notification_runtime.rs
@@ -26,10 +26,6 @@ struct NotificationRuntimeInner {
     // behind one mutex paired with the condvar below.
     state: Mutex<NotificationState>,
     wake: Condvar,
-    // The worker join handle is handed off exactly once between `start()` and
-    // `shutdown()`, so a separate mutex keeps that exclusive ownership transfer
-    // explicit without widening the main state critical section.
-    worker: Mutex<Option<JoinHandle<()>>>,
     path_factory: NotificationPathFactory,
     queue_capacity: usize,
 }
@@ -40,6 +36,7 @@ struct NotificationState {
     shutdown: bool,
     degraded_message: Option<String>,
     queue: VecDeque<NotificationEvent>,
+    worker: Option<JoinHandle<()>>,
 }
 
 impl NotificationRuntime {
@@ -55,7 +52,6 @@ impl NotificationRuntime {
             inner: Arc::new(NotificationRuntimeInner {
                 state: Mutex::new(NotificationState::default()),
                 wake: Condvar::new(),
-                worker: Mutex::new(None),
                 path_factory,
                 queue_capacity,
             }),
@@ -81,34 +77,35 @@ impl NotificationRuntime {
                 AtmError::daemon_unavailable("failed to spawn notification runtime worker")
                     .with_source(source)
             })?;
-        let mut worker = match self.inner.worker.lock() {
-            Ok(worker) => worker,
+        let mut state = match self.inner.state.lock() {
+            Ok(state) => state,
             Err(_) => {
                 let _ = handle.join();
                 return Err(AtmError::daemon_unavailable(
-                    "notification runtime worker lock poisoned",
+                    "notification runtime state lock poisoned",
+                )
+                .with_recovery(
+                    "Restart the daemon; notification lifecycle state can no longer be trusted.",
                 ));
             }
         };
-        *worker = Some(handle);
+        state.worker = Some(handle);
         Ok(())
     }
 
     pub(crate) fn shutdown(&self) -> Result<(), AtmError> {
-        {
+        let handle = {
             let mut state = self.inner.state.lock().map_err(|_| {
                 AtmError::daemon_unavailable("notification runtime state lock poisoned")
+                    .with_recovery(
+                        "Restart the daemon; notification lifecycle state can no longer be trusted.",
+                    )
             })?;
             state.shutdown = true;
             self.inner.wake.notify_all();
-        }
-        if let Some(handle) = self
-            .inner
-            .worker
-            .lock()
-            .map_err(|_| AtmError::daemon_unavailable("notification runtime worker lock poisoned"))?
-            .take()
-        {
+            state.worker.take()
+        };
+        if let Some(handle) = handle {
             let (result_tx, result_rx) = std::sync::mpsc::sync_channel(1);
             std::thread::spawn(move || {
                 let _ = result_tx.send(handle.join());
